Minor tweaks #29

+++ profiles/views.py
@@ -3,6 +3,8 @@ from django.contrib.auth.models import User
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+
+from sellpoint import settings
 from .forms import UserRegisterForm
 
 
@@ -35,15 +37,25 @@ def profile_confirm_delete(request):
     return render(request, 'profiles/profile_confirm_delete.html')
 
 
-@login_required
-def delete_user(request):
+def profile_delete(request):
+    """
+    Deletes profile, which cascades to user.
+    No login_required annotation is needed because
+    the check is performed in the function. This
+    is to avoid the user being redirected to
+    /profiles/delete when logging in after
+    trying to access the url when logged out.
+    """
     context = {}
+
+    if not request.user.is_authenticated:
+        return redirect(settings.LOGIN_URL)
 
     try:
         u = request.user
         logout(request)
         u.delete()
-        messages.success(request, f'Bruker er nå slettet!')
+        messages.success(request, f'Din bruker er nå slettet!')
     except User.DoesNotExist:
         context['msg'] = 'User does not exist.'
     except Exception as e:
